+++ src/main/java/com/timora/app/repository/ProveedorRepository.java
@@ -4,9 +4,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ProveedorRepository extends JpaRepository<Proveedor, Integer> {
+public interface ProveedorRepository extends JpaRepository<Proveedor, Long> {
 
-    Optional<Proveedor> findByUsuario_IdUsuario(Integer idUsuario);
+    Optional<Proveedor> findByUsuario_IdUsuario(Long idUsuario);
 
     Optional<Proveedor> findByNombreNegocio(String nombreNegocio);
 }
